+++ src/test/groovy/dridco/seleniumhtmltojava/commands/WaitForVisibleTest.groovy
@@ -1,0 +1,22 @@
+package dridco.seleniumhtmltojava.commands
+
+class WaitForVisibleTest extends AbstractCommandTest {
+
+	@Override
+	def htmlInstructions() {
+		"""
+<tr>
+        <td>waitForVisible</td>
+        <td>an_element</td>
+        <td></td>
+</tr>
+"""
+		
+	}
+
+	@Override
+	def expectedResult() {
+		'waitForVisible("an_element", "30000");';
+	}
+
+}